Refactor unit tests and move common methods to top-level of tests
